Fixes flask error due to resource not being listed in app.py. Refactors the 'giant plus' into it's own component
